Support for foot notes in LaTeX documents
Foot notes can be referred by using a number in backets [1] as
in this example. The subsequent text must contain the foot-note text
on a line that starts with the marker as follows.

[1] This is the text of the foot note.

The markers can be reused. GOSH scans the text for the next matching
marker starting from the text position of the reference.

Note that foot notes in other output formats than LaTeX are not yet
supported. Hence, they are not described in the official documentation.