html.gosh: support for cross references
Information about cross-html-file references can be supplied via
the new --html-xrefs <xrefs> argument where <xrefs> is a file. The
file has one line per cross reference that looks as follows
(syntactically, it is a Tcl list):

{reference name} type {target}

The 'reference name' is the name of the chapter, section, or whichever
anchor is used, 'type' is the type of reference (chapter, section,
or subsection), 'target' is the link to be supplied as 'href' attribute.

For generating an 'xrefs' file for a given document, the new gosh style
'html_xrefs.gosh' can be used. It accepts the optional argument
--html-path, which defines a path to be used as prefix for the 'href'
attribute.